scripts/requirements.txt
einops

python-dotenv
psycopg2
torch
pandas
transformers[torch]
numpy

openai
pgvector